scheduler-service/requirements.txt
flask==2.3.3
apscheduler==3.10.4
pymysql==1.1.0
requests==2.31.0
python-dotenv==1.0.0
cryptography==41.0.7
